Make the agent smarter and faster
Make it smarter by updating the heuristic used to determine
the value of a gamestate.

Make it faster by keeping the search tree around between turns.
When an opponent move is made, find that game state in the tree
that was already searched, set that as the head, and continue
searching.  This will allow for previously discovered states
to be carried over between turns, there by makeing it faster.